Trying to convert it to doubles...

+++ cAutoPACK/src/Ingredient.cpp
@@ -36,7 +36,7 @@
 
 #include "Ingredient.h"
 
-void Ingredient::setCount(float volume) {
+void Ingredient::setCount(double volume) {
     double nbr = this->molarity * volume * .000602;// #Mod by Graham 8/18/11
     int nbi = (int) nbr; //ceil, floor ?              #Mod by Graham 8/18/11
     double nbmod;
